# repoguard-backend/src/test/resources/contracts/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RepoGuard Agent API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "paths" : {
...
    "/api/v1/config/integrations/github" : {
      "get" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerGetGithubIntegration",
        "x-controller-operation" : "SystemConfigController#getGithubIntegration",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "GithubIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/GithubIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerUpdateGithubIntegration",
        "x-controller-operation" : "SystemConfigController#updateGithubIntegration",
        "parameters" : [ ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "GithubIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GithubIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "GithubIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/GithubIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/integrations/github/test" : {
      "post" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerTestGithubIntegration",
        "x-controller-operation" : "SystemConfigController#testGithubIntegration",
        "parameters" : [ ],
        "requestBody" : {
          "required" : false,
          "x-java-type" : "GithubIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GithubIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ConnectionTestResultDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ConnectionTestResultDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/integrations/mysql" : {
      "get" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerGetMysqlIntegration",
        "x-controller-operation" : "SystemConfigController#getMysqlIntegration",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ServiceIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ServiceIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerUpdateMysqlIntegration",
        "x-controller-operation" : "SystemConfigController#updateMysqlIntegration",
        "parameters" : [ ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "ServiceIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ServiceIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ServiceIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/integrations/mysql/test" : {
      "post" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerTestMysqlConnection",
        "x-controller-operation" : "SystemConfigController#testMysqlConnection",
        "parameters" : [ ],
        "requestBody" : {
          "required" : false,
          "x-java-type" : "ServiceIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ConnectionTestResultDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ConnectionTestResultDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/integrations/rabbitmq" : {
      "get" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerGetRabbitMqIntegration",
        "x-controller-operation" : "SystemConfigController#getRabbitMqIntegration",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ServiceIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ServiceIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerUpdateRabbitMqIntegration",
        "x-controller-operation" : "SystemConfigController#updateRabbitMqIntegration",
        "parameters" : [ ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "ServiceIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ServiceIntegrationConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ServiceIntegrationConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/integrations/rabbitmq/test" : {
      "post" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerTestRabbitMqConnection",
        "x-controller-operation" : "SystemConfigController#testRabbitMqConnection",
        "parameters" : [ ],
        "requestBody" : {
          "required" : false,
          "x-java-type" : "ServiceIntegrationConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceIntegrationConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ConnectionTestResultDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ConnectionTestResultDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/config/review-policy" : {
      "get" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerGetReviewPolicy",
        "x-controller-operation" : "SystemConfigController#getReviewPolicy",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ReviewPolicyConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ReviewPolicyConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerUpdateReviewPolicy",
        "x-controller-operation" : "SystemConfigController#updateReviewPolicy",
        "parameters" : [ ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "ReviewPolicyConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewPolicyConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ReviewPolicyConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ReviewPolicyConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/review-policy/test" : {
      "post" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerTestReviewPolicy",
        "x-controller-operation" : "SystemConfigController#testReviewPolicy",
        "parameters" : [ ],
        "requestBody" : {
          "required" : false,
          "x-java-type" : "ReviewPolicyConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewPolicyConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ConnectionTestResultDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ConnectionTestResultDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/review-rules" : {
      "get" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerGetReviewRules",
        "x-controller-operation" : "SystemConfigController#getReviewRules",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ReviewRulesResponse",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ReviewRulesResponse"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerCreateReviewRule",
        "x-controller-operation" : "SystemConfigController#createReviewRule",
        "parameters" : [ ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "ReviewRuleConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewRuleConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ReviewRuleConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ReviewRuleConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/review-rules/{id}" : {
      "put" : {
        "tags" : [ "SystemConfig" ],
        "operationId" : "systemConfigControllerUpdateReviewRule",
        "x-controller-operation" : "SystemConfigController#updateReviewRule",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 64
          }
        } ],
        "requestBody" : {
          "required" : true,
          "x-java-type" : "ReviewRuleConfigRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewRuleConfigRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "x-java-response-envelope" : "ApiResponse",
            "x-java-response-data" : "ReviewRuleConfigDto",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-java-type" : "ApiResponse",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "code" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "data" : {
                      "$ref" : "#/components/schemas/ReviewRuleConfigDto"
                    },
                    "timestamp" : {
                      "type" : "string",
                      "format" : "date-time"
                    }
                  },
                  "required" : [ "success", "code", "message", "timestamp" ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GithubIntegrationConfigDto" : {
        "type" : "object",
        "x-java-type" : "GithubIntegrationConfigDto",
        "properties" : {
          "provider" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "baseUrl" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "defaultOwner" : {
            "type" : "string"
          },
          "defaultRepo" : {
            "type" : "string"
          },
          "lastCheckedAt" : {
            "type" : "string"
          },
          "lastError" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          },
          "secretStatus" : {
            "type" : "string"
          }
        }
      },
      "GithubIntegrationConfigRequest" : {
        "type" : "object",
        "x-java-type" : "GithubIntegrationConfigRequest",
        "properties" : {
          "baseUrl" : {
            "type" : "string",
            "minLength" : 1,
            "maxLength" : 512
          },
          "token" : {
            "type" : "string",
            "maxLength" : 4096
          },
          "defaultOwner" : {
            "type" : "string",
            "maxLength" : 128
          },
          "defaultRepo" : {
            "type" : "string",
            "maxLength" : 128
          }
        },
        "required" : [ "baseUrl" ]
      },
...
      "ServiceIntegrationConfigRequest" : {
        "type" : "object",
        "x-java-type" : "ServiceIntegrationConfigRequest",
        "properties" : {
          "baseUrl" : {
            "type" : "string",
            "minLength" : 1,
            "maxLength" : 512
          },
          "username" : {
            "type" : "string",
            "maxLength" : 128
          },
          "secret" : {
            "type" : "string",
            "maxLength" : 4096
          },
          "resource" : {
            "type" : "string",
            "maxLength" : 128
          }
        },
        "required" : [ "baseUrl" ]
      },
...
    }
  }
}